requirements.txt
bcrypt==4.0.1
blinker==1.6.2
click==8.1.7
colorama==0.4.6
flask==2.2.3
Flask-Bcrypt==1.0.1
Flask-JWT-Extended==4.5.2
flask-marshmallow==0.15.0
flask-sqlalchemy==3.0.3
greenlet==2.0.2
gunicorn==21.2.0
importlib-metadata==6.7.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
marshmallow==3.19.0
marshmallow-sqlalchemy==0.29.0
packaging==23.1
psycopg2-binary==2.9.7
PyJWT==2.8.0
python-dotenv==0.21.1
SQLAlchemy==2.0.20
typing-extensions==4.7.1
werkzeug==2.2.3
zipp==3.15.0
